# zend_compile: Add support for `%d` to `sprintf()` optimization