--- a/ext/standard/tests/strings/sprintf_rope_optimization_004.phpt
+++ b/ext/standard/tests/strings/sprintf_rope_optimization_004.phpt
@@ -10,12 +10,12 @@ $b = new GMP("-1337");
 $c = new GMP("999999999999999999999999999999999");
 
 try {
-	var_dump(sprintf("%d/%d/%d", $a, $b, $c));
+	var_dump(sprintf("%d/%d/%d/%s", $a, $b, $c, $c + 1));
 } catch (\Throwable $e) {echo $e, PHP_EOL; } echo PHP_EOL;
 
 echo "Done";
 ?>
 --EXPECTF--
-string(28) "42/-1337/4089650035136921599"
+string(63) "42/-1337/4089650035136921599/1000000000000000000000000000000000"
 
 Done